refactor(auth): reorganize API imports and move auth/user functions to separate files
